refactor, remove dead code
