+++ dascore/utils/misc.py
@@ -11,7 +11,7 @@ import numpy as np
 import numpy.typing as nt
 import pandas as pd
 
-from dascore.exceptions import ParameterError, PatchDimError
+from dascore.exceptions import ParameterError
 
 
 def register_func(list_or_dict: Union[list, dict], key: Optional[str] = None):
@@ -148,29 +148,6 @@ def _get_sampling_rate(sampling_period):
     else:
         num = 1
     return num / sampling_period
-
-
-def get_dim_value_from_kwargs(patch, kwargs):
-    """
-    Assert that kwargs contain one value and it is a dimension of patch.
-
-    Several patch functions allow passing values via kwargs which are dimension
-    specific. This function allows for some sane validation of such functions.
-
-    Return the name of the dimension, its axis position, and its value.
-    """
-    dims = patch.dims
-    overlap = set(dims) & set(kwargs)
-    if len(kwargs) != 1 or not overlap:
-        msg = (
-            "You must use exactly one dimension name in kwargs. "
-            f"You passed the following kwargs: {kwargs} to a patch with "
-            f"dimensions {patch.dims}"
-        )
-        raise PatchDimError(msg)
-    dim = list(overlap)[0]
-    axis = dims.index(dim)
-    return dim, axis, kwargs[dim]
 
 
 def all_close(ar1, ar2):
