+++ dascore/utils/patch.py
@@ -5,7 +5,6 @@ import collections
 import copy
 import functools
 import inspect
-import itertools
 import re
 from fnmatch import translate
 from typing import (
@@ -607,29 +606,24 @@ def get_default_patch_name(patch):
     return f"DAS__{net}__{sta}__{tag}__{start}__{end}"
 
 
-def get_dim_coords(coords, dims=None):
+def get_dim_value_from_kwargs(patch, kwargs):
     """
-    Get only coords that correspond to dimensions. Return dict.
+    Assert that kwargs contain one value and it is a dimension of patch.
+
+    Several patch functions allow passing values via kwargs which are dimension
+    specific. This function allows for some sane validation of such functions.
+
+    Return the name of the dimension, its axis position, and its value.
     """
-
-    def _strip_dimension_names(value):
-        """Strip out dimension names."""
-        if len(value) and isinstance(value[0], (str, tuple, list)):
-            assert len(value) >= 2
-            return value[1]
-
-    def _get_dimensional_names(coords):
-        """Infer dimensional names"""
-        candidates = set(coords)
-        # get the coordinates that can be listed as the first of a tuple
-        nested_coord_names = [
-            v[0] if len(v) and isinstance(v[0], (str, tuple, list)) else i
-            for i, v in coords.items()
-        ]
-        dim_set = set(itertools.chain(nested_coord_names)) & candidates
-        # This tries to preserve the order in coords
-        return [x for x in coords if x in dim_set]
-
-    dims = dims or _get_dimensional_names(coords)
-    sub = {x: _strip_dimension_names(coords[x]) for x in dims}
-    return sub
+    dims = patch.dims
+    overlap = set(dims) & set(kwargs)
+    if len(kwargs) != 1 or not overlap:
+        msg = (
+            "You must use exactly one dimension name in kwargs. "
+            f"You passed the following kwargs: {kwargs} to a patch with "
+            f"dimensions {patch.dims}"
+        )
+        raise PatchDimError(msg)
+    dim = list(overlap)[0]
+    axis = dims.index(dim)
+    return dim, axis, kwargs[dim]
